backend/requirements.txt
fastapi==0.115.0
psycopg2-binary>=2.9
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
supabase==2.7.4
anthropic==0.34.2
openai==1.45.0
youtube-transcript-api==1.2.4
google-api-python-client==2.143.0
resend==2.4.0
numpy==1.26.4
scikit-learn==1.5.1
httpx==0.27.2
firecrawl-py==1.15.0
python-multipart==0.0.9
PyJWT[cryptography]>=2.8
pytest>=8.0
pytest-asyncio>=0.23
